Extract struct parsing to another file
